Moved a few discord-related functions to common/discord.py.
